adjusted math for calculating the ship materials, appears we're median of three other sources, moved the bill of materials to its own component and placed it back in the center, ship description on the right
